move tip_tracker and well_grid to geometry.
