+++ api/tests/opentrons/protocols/geometry/test_tip_tracker.py
@@ -1,9 +1,9 @@
 from typing import List
 
 import pytest
-from opentrons.protocols.context.tip_tracker import TipTracker
+from opentrons.protocols.geometry.tip_tracker import TipTracker
 from opentrons.protocols.context.well import WellImplementation
-from opentrons.protocols.context.well_grid import WellGrid
+from opentrons.protocols.geometry.well_grid import WellGrid
 
 
 @pytest.fixture()
